# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion de Comptes - Banque",
        "description": "API REST pour la gestion des comptes bancaires avec authentification OAuth2",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.banque.example.com/api/v1",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/v1/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Lister tous les clients",
                "description": "Récupère la liste paginée des clients selon les permissions de l'utilisateur. Admin voit tous les clients, Client voit uniquement son propre profil.",
                "operationId": "getClients",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par nom, prénom ou email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "nom",
                                "prenom",
                                "email",
                                "dateCreation"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des clients récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Clients récupérés avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Client"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "currentPage": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "totalPages": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "totalItems": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "itemsPerPage": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "hasNext": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "hasPrevious": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "properties": {
                                                "self": {
                                                    "type": "string",
                                                    "example": "/api/v1/clients?page=1&limit=10"
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "/api/v1/clients?page=2&limit=10"
                                                },
                                                "first": {
                                                    "type": "string",
                                                    "example": "/api/v1/clients?page=1&limit=10"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "/api/v1/clients?page=3&limit=10"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "read-clients"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "title": "Client",
                "description": "Représentation d'un client bancaire",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Diallo"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Amadou"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "amadou.diallo@example.com"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+221771234567"
                    },
                    "adresse": {
                        "type": "string",
                        "example": "Dakar, Sénégal"
                    },
                    "dateNaissance": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-15"
                    },
                    "dateCreation": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-03-15T00:00:00Z"
                    },
                    "comptes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Compte"
                        }
                    },
                    "metadata": {
                        "properties": {
                            "derniereModification": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2023-06-10T14:30:00Z"
                            },
                            "nombreComptes": {
                                "type": "integer",
                                "example": 2
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Compte": {
                "title": "Compte",
                "description": "Représentation d'un compte bancaire",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "numeroCompte": {
                        "type": "string",
                        "example": "C00123456"
                    },
                    "titulaire": {
                        "type": "string",
                        "example": "Amadou Diallo"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "epargne",
                            "cheque"
                        ],
                        "example": "epargne"
                    },
                    "solde": {
                        "type": "number",
                        "format": "decimal",
                        "example": 1250000
                    },
                    "devise": {
                        "type": "string",
                        "example": "FCFA"
                    },
                    "dateCreation": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-03-15T00:00:00Z"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "ferme"
                        ],
                        "example": "bloque"
                    },
                    "motifBlocage": {
                        "type": "string",
                        "example": "Inactivité de 30+ jours",
                        "nullable": true
                    },
                    "metadata": {
                        "properties": {
                            "derniereModification": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2023-06-10T14:30:00Z"
                            },
                            "version": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel Passport OAuth2 security",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": {
                            "read-comptes": "Lire les comptes",
                            "write-comptes": "Modifier les comptes",
                            "admin": "Accès administrateur"
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        {
            "passport": [
                "read-comptes"
            ]
        }
    ]
}